requirements.txt
flask==2.2.3
requests==2.28.2
Jinja2==3.1.2
flask_mail==0.9.1
bcrypt==4.0.1
flask_toastr==0.5.8
flask_sqlalchemy
mysql-connector-python
Flask-Uploads
